requirements.txt
yfinance==0.2.36
pandas==2.3.0
python-dotenv==1.2.2
azure-storage-blob==12.28.0
snowflake-connector-python==4.4.0
pyspark==4.1.1
ollama
streamlit